allow recognition of keywords
Keywords look like ordinary identifiers, but we'll want the language to
treat them specially and the lexer should probably take note of them
specially. The reader may have observed that class now called BaseLexer
is more of a namespace for a set of functions with shared "global" state
than a meaningful object-object. But I'm OK with this.