append!
If nothing else, I'm starting to get a better model of the extent and
manner in which the current implementation of Glitteral cheats me out of
the interesting and noble parts of writing a compiler; I've built an AST
of sorts (the Expression objects), but I fully expect that translating
that to Rust source in the naive way, while producing an impressive demo
very quickly, is just not going to scale as the language gets more
features.

For example, Rust has this "semicolons suppress expression values, but
there are times you need the semicolon to separate expressions from
'top-level' syntax" thing going on, which caused problems with the demo
code I used to test this implementation, which I solved for the case I
was considering by plumbing through this `statementlike` optional
argument that is true at the top level but false nested within other
expressions (in order to get a semicolon at the end of my standalone
`append!` call, but none when passing an arithmetic application as a
function value)—but you only have so much complexity budget for ad hoc
hacks to get a specific example to work without understanding the
theory. Similar things could be said about the
`environment`/`builtins`/`represent_identifiable` tangle here not based
on any well-understood theoretical justification.