move slowly and break things
This breaks my demo code, but the demo code didn't really deserve to
work. We only got away with this "if we can't find this name in the
environment, just print the name anyway" sophistry because we've been
compiling ("compiling") to _another high-level language_; if we were
doing the real work of reducing computation, we wouldn't even have the
vocabulary in which to tell this particular lie. And the most obvious
remedy of "just" adding function argument names into the global
environment just like (global) variable definitions will only lead to
grief; we can predict that a merely human compiler author won't have the
extreme mental discipline needed to correctly juggle scopes. What we
need is a clear procedure (an algorithmic lucidity) for producing the
nested environments automatically. What we need is to take a step back
and think carefully about how to do this correctly. What we need is to
annotate the AST, just like the textbook was trying to tell us all
along.