"Be nice, girl"
After introducing a separate function for subscript access (eventually
we want Clojure-style containers-as-functions, but in this sparkling
moment, we have License to Hack; I was going to choose `get` as the
identifier with the honors, before noticing that a lone underscore
better fits the subscript assignment function and the aesthetic of
Glitteral in general) and casting to usize, our aspiration bubblesort
example ("example") compiles! ... but doesn't seem to sort at all, one
notes with disappointment and mild but undeserved surprise—until
noticing that actually calling the `bubblesort` function would probably
help with that. Then we get thread "'<main>' panicked at 'index out of
bounds: the len is 5 but the index is 5'", which is more exciting still.