"They like to give advice, girl"
I was surprised at how smoothly the bubblesort and fizzbuzz examples
went; now, with the slightly greater complexity of a textbooky quicksort
example, we're starting to get a better look at the tension if not sheer
absurdity of trying to reconcile the semantics that I'm ever-so-vaguely
imagining for Glitteral with Rust's binding scope rules and the
widely-execrated and ever-obdurate borrow checker.

It's a start? Along the way (the long, perilous way), we implement
`when` expressions and decide to use the interogative convention for
predicates maybe?