update README: "associative", 3 spaces, forget docstrings, syntax reform
I feel like docstrings make more sense in an interpreted language where
you're going to be typing `(help my_function)` at the REPL, whereas so
far all my attention has been focused on the compilation-via-Rust
backend (but we can imagine that I should write an interpreter backend,
too?).

I feel like the syntax reform is going to be so great when I actually
get to it!