dictionary code generation; fundamental doubts weakened if not dispelled
While I was working on these changes, I found myself wondering about
whether starting Glitteral with a Rust backend with my current knowledge
level was even a good idea in the first place (as opposed to trying a
surmisedly less ambitious interpreter written in Python while shoring up
my textbook knowledge), the fear being that the current approach could
perhaps only lead to an unmaintainable, incomprehensible mess of code
that only works for the particular special cases that I considered and
_made_ it work for (whereas if your design is backed by a good theory,
the theory can guarantee correct behavior that even the designer may not
have anticipated).

And, well ... that still might be actually true in the long run. But
seeing this correctly populate a Rust HashMap (albeit with no way to
access it from Glitteral as yet) gives me hope that the attempt isn't a
waste of time; I'm learning a lot and am on track to make something
really impressive on the timescale of weeks or months ("impressive" in
the sense of credibly certifying that this gloriously unschooled junior
engineer is making credible strides towards mastering their chosen
profession through proper study and practice, not "impressive" like
something you'd actually want to use).