rescue expected output test from cruelly overzealous .gitignore
This really should have been committed with 3a309a5c, but thanks to
http://stackoverflow.com/a/19023985, I had a clever ("clever")
.gitignore in the eg/ directory that was intended to ignore files
lacking an extension (so that I could keep with the tradition of
binaries not having extensions, without accidentally checking in my
binaries; one imagines that cargo solves this with the separate target/
directory). Unfortunately, this seems to have ignored the entire tests
subdirectory; no longer trusting cleverness, I may be doomed to have to
update the .gitignore as I add new examples (which doesn't mean manual
labor, of course!—we could make a Git hook or Invoke task).