early rising action towards syntax reform
You know, I had wondered if I was wrong about making a Lexer class when
I was only really using the `lex` function that wraps it—I told myself
it was probably only for the convenience of being able to pull out
functions with shared state instead of explicitly passing the state
around as arguments. But the actual use of the Lexer objects in
these (skipped for only another moment in the timeless history of the
commit graph) tests reassures me that my initial design intuition was
right in ways that I couldn't have articulated at the time: at least now
that I've moved some setup from `tokenize` to the __init__ializer, a
Lexer can tokenize some source, have its persistent state inspected, and
then tokenize some more; it really deserves to be an object.

But mostly this commit sketches an unproven but promising-seeming design
for how to accomodate Indent and Dedent tokens within our current lexing
scheme; it seems worth while to check in these foundations before
tearing into and rearranging the soul of the tokenizer logic itself.