syntax reform rising action: the First Battle of Lexington
The "strip trailing newlines when recognizing a token" workaround that I
had improvised in response to the unfortunate semantics of '$' in
regexes ("or just before the newline at the end of the string", spare
me! Whatever happened to "Special cases aren't special enough to break
the rules"?!) won't suffice if we want to recognize things like '\n' as
Indent tokens (in the right context). But we can slap a negative
lookahead assertion on everything in sight.